sqrily/requirements.txt
# Essential FastAPI stack
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
itsdangerous>=2.1.0

# OAuth integrations
authlib>=1.2.0
httpx>=0.25.0

# AI & Machine Learning
openai>=1.40.0
tiktoken>=0.5.0  # Token counting for OpenAI
tenacity>=8.2.0  # Retry logic for API calls
aiofiles>=23.2.0  # Async file operations for AI

# WebSocket support
websockets>=11.0.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Monitoring & Logging
prometheus-client>=0.18.0
structlog>=23.1.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3